backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yfinance==0.2.28
supabase==2.15.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx>=0.26,<0.29
pandas==2.1.4
numpy==1.26.2
asyncio==3.4.3
schedule==1.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
websockets==12.0
redis==5.0.1
celery==5.3.4 